% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastaTools.R
\name{concatenateFastas}
\alias{concatenateFastas}
\title{Concatenate all fastas in a Folder.}
\usage{
concatenateFastas(path, pattern = "\\\\.fa", outfile = "combined.fasta")
}
\arguments{
\item{path}{Path to file containing folder}

\item{pattern}{Pattern used to grep filenames (default: ".fa")}

\item{outfile}{Name of the combined file.}
}
\details{
Can be used to produce a single Fasta ready for library generation.
File is saved in the path directory.
}

