% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadAnnotations.R
\name{downloadAnnotations}
\alias{downloadAnnotations}
\title{Download transcriptome annotatin from Ensemblin a format that customProDB can use
This Uses the import function from CustomProDB in combination with BioMart and can take a few hours.}
\usage{
downloadAnnotations(annotation_path, archive = FALSE,
  host = "dec2016.archive.ensembl.org")
}
\arguments{
\item{annotation_path}{Path to a folder where Annotation files should be stored.}
}
\description{
Download transcriptome annotatin from Ensemblin a format that customProDB can use
This Uses the import function from CustomProDB in combination with BioMart and can take a few hours.
}

