% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPKMTableTools.R
\name{filterDetectedTranscripts}
\alias{filterDetectedTranscripts}
\title{Filter FPKM Table}
\usage{
filterDetectedTranscripts(fpkm_table, quiet = FALSE)
}
\arguments{
\item{fpkm_table}{Table with FPKM vaules of (multiple) samples}

\item{quiet}{print how many rows are filtered, default FALSE}
}
\description{
Filter FPKM Table
}
\details{
Filters out all rows where transcripts are detected in every sample
}

