% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionTools.R
\name{filterSpacePlot}
\alias{filterSpacePlot}
\title{Generate a Junction Summary Report}
\usage{
filterSpacePlot(covfilter_unique = c(1, 2, 3, 4, 5, 10), share_num = c(1,
  2), bed_directory, splicemax, txdb, ids)
}
\arguments{
\item{splicemax}{Splice Matrix from annotation folder}

\item{txdb}{Txdb object from annotation folder}

\item{ids}{Id mapping from annotation folder}

\item{sharedjun}{GRanges object containing all splice junctions}

\item{print}{boolean, should the result be printed in working directory as pdf
or be returned as a list.}
}
\description{
Generate a Junction Summary Report
}
\details{
Saves a table of the types of the Splice junctions to working directory
}
