% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastaTools.R
\name{finishFasta}
\alias{finishFasta}
\title{Finish a raw fasta file for analysis}
\usage{
finishFasta(file, outfile = paste0(file, "iRT_cRAP.fasta"), removeStars = T,
  addcRAP = TRUE, addiRT = TRUE)
}
\arguments{
\item{file}{PName of the input file}

\item{outfile}{Name of the combined file.}

\item{removeStars}{logical, Whether to remove the stars at sequence ends}

\item{addcRAP}{logical, Should cRAP and affinity tag sequences be appended}

\item{addiRT}{logical, Should iRT peptides be appended}
}
\details{
Can be used to produce a single Fasta ready for library generation.
}
