% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPKMTableTools.R
\name{FPKMhist}
\alias{FPKMhist}
\title{Output Histogram of FPKM values}
\usage{
FPKMhist(fpkm_table, qt)
}
\arguments{
\item{fpkm_table}{Table with FPKM vaules of (multiple) samples}

\item{mapping_table}{table with ids of both types}
}
\description{
Output Histogram of FPKM values
}
\details{
Plots a Histogram with the 30% quantile as a vertical line
}

