% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPKMTableTools.R
\name{FPKMhistProteincoding}
\alias{FPKMhistProteincoding}
\title{Output Histogram of FPKM values for protein-coding vs. non protein coding}
\usage{
FPKMhistProteincoding(fpkm_table, id_table, qt)
}
\arguments{
\item{fpkm_table}{Table with FPKM vaules of (multiple) samples}

\item{id_table}{table with ids of both types}
}
\description{
Output Histogram of FPKM values for protein-coding vs. non protein coding
}
\details{
Plots a Histogram with the 30% quantile as a vertical line
}

