% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPKMTableTools.R
\name{FPKMsummaryReport}
\alias{FPKMsummaryReport}
\title{Print Pdf with plots to help evaluate FPKM distribution}
\usage{
FPKMsummaryReport(fpkm_table, id_table, qt = c(0.2, 0.22, 0.25, 0.3),
  print = TRUE)
}
\arguments{
\item{fpkm_table}{Table with FPKM vaules of (multiple) samples}

\item{id_table}{table with ids of both types}

\item{print}{boolean, should the result be printed in working directory as pdf
or returned as list of ggplot objects}
}
\description{
Print Pdf with plots to help evaluate FPKM distribution
}
\details{
Plots a Histograms and sd vs. FPKM
}

