% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputJunctions.R
\name{getAltSpliceProteins}
\alias{getAltSpliceProteins}
\title{Get alternatively spliced proteins}
\usage{
getAltSpliceProteins(jun, exon, procodingseq)
}
\arguments{
\item{jun}{Junction file produced by the JunctionType function}

\item{procodingseq}{Protein coding sequences from annotation folder}

\item{exons}{Exon table from annotation folder}
}
\description{
Uses junction table to produce spliced sequences of known transcripts. A protein is produced when a
junction connects two of its exons that are not adjacent.
}

