% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastaTools.R
\name{getBioMart}
\alias{getBioMart}
\title{Download uniprot ids from BioMart}
\usage{
getBioMart(ids)
}
\arguments{
\item{ids}{Id mapping from annotation folder}
}
\description{
Download uniprot ids from BioMart
}
\details{
internal function
}
