% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCufflinksFPKM.R
\name{getCufflinksFPKM}
\alias{getCufflinksFPKM}
\title{Import FPKM from Cufflinks}
\usage{
getCufflinksFPKM(files, filetype = c("isoforms.fpkm_tracking",
  "genes.fpkm_tracking", "GTF"))
}
\arguments{
\item{files}{Path to Cufflinks files to import. Can be isoforms.fpkm_tracking, genes.fpkm_tracking, or GTF}

\item{filetype}{possible values : "isoforms.fpkm_tracking", "genes.fpkm_tracking", "GTF"}
}
\description{
Takes multiple Cufflinks output files as input and outputs a data.frame with FPKM values
}

