% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCufflinksFPKM.R
\name{getCufflinksgeneFPKM}
\alias{getCufflinksgeneFPKM}
\title{Takes Cufflinks genes.fpkm_tracking output files as input and outputs a named numeric vector (names = Transcript/gene id, value = FPKM)}
\usage{
getCufflinksgeneFPKM(genes.fpkm_tracking_file_path)
}
\arguments{
\item{genes.fpkm_tracking_file_path}{Path to GTF file that is output by cufflinks}
}
\description{
Takes Cufflinks genes.fpkm_tracking output files as input and outputs a named numeric vector (names = Transcript/gene id, value = FPKM)
}
