% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPKMTableTools.R
\name{getProteinCoding}
\alias{getProteinCoding}
\title{Convert FPKM Table with transcript ids to protein ids}
\usage{
getProteinCoding(fpkm_table, id_table, transcript_id_name = "tx_name",
  protein_id_name = "pro_name")
}
\arguments{
\item{fpkm_table}{Table with FPKM vaules of (multiple) samples}

\item{id_table}{table with ids of both types}

\item{transcript_id_name}{colname of the transcript id column in mapping table}

\item{protein_id_name}{colname of the protein id column in mapping table}
}
\description{
Convert FPKM Table with transcript ids to protein ids
}
\details{
Matches rownames of fpkm table to protein ids with a mapping table (has to be supplied)

Rows with no matching protein id are discarded
}
