% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareVariants.R
\name{getSNVandINDEL}
\alias{getSNVandINDEL}
\title{Annotate Variants}
\usage{
getSNVandINDEL(granges_variants, exon_annotation, procodingseq)
}
\arguments{
\item{granges_variants}{GRanges object produced by getVcfs}

\item{exon_annotation}{Exon annotation from annotation folder}

\item{procodingseq}{Coding sequences from annotation folder}
}
\description{
Annotate Variants
}
\details{
Annotates the variants in a granges object produced by getVcfs
}

