% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStarJunctions.R
\name{getStarJunctions}
\alias{getStarJunctions}
\title{Read in STAR junctions}
\usage{
getStarJunctions(SJout_dir, splicemax, txdb, ids, pattern = "\\\\.tab",
  share_num = 1, extend = 100, skip = 0, covfilter_unique = 1,
  covfilter_multi = 0, overhang = "max_overhang")
}
\arguments{
\item{SJout_dir}{Path to Star output files}
}
\description{
Takes STAR output files (SJ.out.tab as input and outputs a
named numeric vector (names = Transcript/gene id, value = FPKM)
}

