% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareVariants.R
\name{getVcfs}
\alias{getVcfs}
\title{Read Vcf files to GRanges}
\usage{
getVcfs(vcf_dir, pattern = ".vcf", share_num = "100\%")
}
\arguments{
\item{vcf_dir}{Path to directory with Vcf files of all replicate samples}

\item{pattern}{Pattern to grep the specific Vcf files in the directory}

\item{share_num}{Number of samples that must contain the variant.
Two options, percentage format or sample number.}
}
\description{
Read Vcf files to GRanges
}
\details{
Takes VCF output files from variant calling pipeline and combines
 a GRanges object ready for fasta generation.
}
