% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionTools.R
\name{junctionSummaryReport}
\alias{junctionSummaryReport}
\title{Plot the distribution of your splice junction types}
\usage{
junctionSummaryReport(junction_type, pdf = TRUE,
  outfile = "Junction_summary_report.pdf")
}
\arguments{
\item{junction_type}{generated by the function: JunctionType}

\item{pdf}{boolean, should the result be printed in working directory as pdf
or be returned as a list.}

\item{outfile}{Name of the printed pdf}
}
\description{
Plot the distribution of your splice junction types
}
