% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPKMTableTools.R
\name{makeDict}
\alias{makeDict}
\title{internal, makes named vector dict}
\usage{
makeDict(ids, mapping_table, query_id_type, map_id_type)
}
\arguments{
\item{ids}{query ids}

\item{mapping_table}{table with ids of both types}

\item{query_id_type}{name of the query column}

\item{map_id_type}{name of the map_id column}
}
\description{
internal, makes named vector dict
}

