% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputJunctions.R
\name{OutputJunctions}
\alias{OutputJunctions}
\title{Output Splice Junction FASTA}
\usage{
OutputJunctions(outfile, juncs, exons, procodingseq, proteinseq, genome,
  alt_spl = TRUE, all_junc_peptides = TRUE, trimTry = TRUE)
}
\arguments{
\item{exons}{Exon table from annotation folder}

\item{procodingseq}{Protein coding sequences from annotation folder}

\item{jun}{Junction file produced by the JunctionType function}
}
\description{
Takes STAR output files (SJ.out.tab as input and outputs a
named numeric vector (names = Transcript/gene id, value = FPKM)
}

