% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputJunctions.R
\name{OutputNovelJun}
\alias{OutputNovelJun}
\title{Generate peptide FASTA file that contains novel junctions.}
\usage{
OutputNovelJun(junction_type, genome, trimTry = F, proteinseq, ...)
}
\arguments{
\item{junction_type}{a data frame which is the output of function JunctionType()}

\item{genome}{a BSgenome object. (e.g. Hsapiens)}

\item{proteinseq}{a data frame cotaining amino acid sequence for each protein.}

\item{...}{Additional arguments.}

\item{outfile}{output file name}
}
\value{
FASTA file that contains novel junction peptides.
}
\description{
Three-frame translation of novel junctions. And remove those could be found in normal protein sequences.
This function requires a genome built by BSgenome package.
}
\author{
Xiaojing Wang

Max Frank
}

