% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionTools.R
\name{plotJuncReplicateOverlap}
\alias{plotJuncReplicateOverlap}
\title{#' Generate a Junction Summary Report
#' @details Saves a table of the types of the Splice junctions to working directory
#' @param junction_type generated by the function: JunctionType
#' @export}
\usage{
plotJuncReplicateOverlap(bed_directory, pattern = "\\\\.tab",
  outfile = "SJ_Replicate_overlap.tiff")
}
\arguments{
\item{bed_directory}{Path to folder with all replicate vcf files}

\item{pattern}{pattern to match}
}
