% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPKMTableTools.R
\name{plotSDvsMeanExpression}
\alias{plotSDvsMeanExpression}
\title{Scatterplot of sd vs log transformed expression}
\usage{
plotSDvsMeanExpression(fpkm_table)
}
\arguments{
\item{fpkm_table}{Table with FPKM vaules of (multiple) samples}
}
\description{
Scatterplot of sd vs log transformed expression
}
\details{
Log transforms expression, calculates sd and plots it over expression dimension
}
