% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareVariants.R
\name{plotSNVReplicateOverlap}
\alias{plotSNVReplicateOverlap}
\title{Plot the overlap of identified Variants between replicates}
\usage{
plotSNVReplicateOverlap(vcf_dir = vcf_directory, pattern = ".vcf$",
  outfile = "SNVReplicate_overlap.tiff")
}
\arguments{
\item{vcf_dir}{Path to folder with all replicate vcf files}

\item{pattern}{pattern to match}
}
\description{
Plot the overlap of identified Variants between replicates
}
