% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareVariants.R
\name{VariantSummaryReport}
\alias{VariantSummaryReport}
\title{Generate a Variant Summary Report}
\usage{
VariantSummaryReport(snv_and_indel, ids, txdb, print = TRUE)
}
\arguments{
\item{snv_and_indel}{List object produced by getSNVandINDEL}

\item{ids}{Id mapping from annotation folder}

\item{txdb}{Txdb object from annotation folder}

\item{print}{boolean, should the result be printed in working directory as pdf
or be returned as a list.}
}
\description{
Generate a Variant Summary Report
}
\details{
Prints a table of the location of the variants. Caution:
Because the Varlocation function of CustomProDB is very slow, this takes a long time
to compute.
}

