% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastaTools.R
\name{concatenateFastas}
\alias{concatenateFastas}
\title{Concatenate all fastas in a Folder.}
\usage{
concatenateFastas(path, pattern = "\\\\.fa", outfile = "combined.fasta",
  addcRAP = TRUE, addiRT = TRUE, saveMappingTable = TRUE)
}
\arguments{
\item{path}{Path to file containing folder}

\item{pattern}{Pattern used to grep filenames (default: ".fa")}

\item{outfile}{Name of the combined file.}

\item{addcRAP}{logical, Should cRAP and affinity tag sequences be appended}

\item{addiRT}{logical, Should iRT peptides be appended}
}
\details{
Can be used to produce a single Fasta ready for library generation.
File is saved in the path directory.
}
