% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCufflinksFPKM.R
\name{getCufflinksFPKM}
\alias{getCufflinksFPKM}
\title{Takes Cufflinks gtf output files as input and outputs a named numeric vector (names = Transcript/gene id, value = FPKM)}
\usage{
getCufflinksFPKM(gtf_file_path)
}
\arguments{
\item{gtf_file_path}{Path to GTF file that is output by cufflinks}
}
\description{
Takes Cufflinks gtf output files as input and outputs a named numeric vector (names = Transcript/gene id, value = FPKM)
}

